// src/lib/components/Image/image.d.ts
export type LayoutValue = 'fill' | 'fixed' | 'intrinsic' | 'responsive';
export type ObjectFitStyle = 'contain' | 'cover' | 'fill' | 'none' | 'scale-down';
export type ObjectPositionStyle = CSSStyleDeclaration['objectPosition'];
export type PlaceholderValue = 'blur' | 'empty';
export type OnLoadingComplete = (result: { naturalWidth: number; naturalHeight: number }) => void;